refactor(): separate ieBuilder, peBuilder & networkIssueDeterctor from EventMapper
